+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - Fix Atlas Scientific UART interfaces
  - Fix display of units in conversion list on Measurement Settings page
  - Fix unit conversions for Math controllers ([#716](https://github.com/kizniche/mycodo/issues/716))
+ - Fix Wet-Bulb Humidity calculation in Math controller ([#716](https://github.com/kizniche/mycodo/issues/716))
 
 ### Features